Move fixture code into more specific package - #19867
