+++ subprojects/kotlin-dsl-tooling-builders/src/crossVersionTest/groovy/org/gradle/kotlin/dsl/tooling/builders/r60/AbstractKotlinDslScriptsModelCrossVersionSpec.groovy
@@ -16,8 +16,8 @@
 
 package org.gradle.kotlin.dsl.tooling.builders.r60
 
+import org.gradle.integtests.fixtures.build.BuildSpec
 import org.gradle.kotlin.dsl.tooling.builders.AbstractKotlinScriptModelCrossVersionTest
-import org.gradle.integtests.fixtures.BuildSpec
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.tooling.model.kotlin.dsl.KotlinDslScriptsModel
