+++ subprojects/kotlin-dsl-tooling-builders/src/crossVersionTest/groovy/org/gradle/kotlin/dsl/tooling/builders/r68/AbstractKotlinDslScriptsModelCrossVersionSpec.groovy
@@ -16,8 +16,8 @@
 
 package org.gradle.kotlin.dsl.tooling.builders.r68
 
+import org.gradle.integtests.fixtures.build.BuildSpec
 import org.gradle.kotlin.dsl.tooling.builders.AbstractKotlinScriptModelCrossVersionTest
-import org.gradle.integtests.fixtures.BuildSpec
 import org.gradle.tooling.model.kotlin.dsl.KotlinDslScriptsModel
 
 import static org.gradle.integtests.tooling.fixture.TextUtil.escapeString
